Arreglados errores y añadidos estilos

+++ src/GFit-Analysis-Back/src/main/java/com/gfitanalysis/GFitAnalysisBack/repository/UserRepositoryI.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import com.gfitanalysis.GFitAnalysisBack.model.User;
 
 @Repository
-public interface UserRepositoryI extends JpaRepository<User, Long> {
+public interface UserRepositoryI extends JpaRepository<User, Integer> {
 	
 	public User findByEmail(String email);
